runtime: moving code around
