+++ source/common/http/http2/codec_impl.cc
@@ -20,7 +20,6 @@
 #include "common/http/header_utility.h"
 #include "common/http/headers.h"
 #include "common/http/utility.h"
-#include "common/runtime/runtime_impl.h"
 
 #include "absl/container/fixed_array.h"
