+++ source/common/http/http2/conn_pool.cc
@@ -7,6 +7,7 @@
 
 #include "common/http/http2/codec_impl.h"
 #include "common/http/http2/conn_pool_legacy.h"
+#include "common/runtime/runtime_features.h"
 
 namespace Envoy {
 namespace Http {
