+++ source/common/runtime/BUILD
@@ -17,10 +17,26 @@ envoy_cc_library(
         "runtime_features.h",
     ],
     deps = [
+        # AVOID ADDING TO THESE DEPENDENCIES IF POSSIBLE
+        # Any code using runtime guards depends on this library, and the more dependencies there are,
+        # the harder it is to runtime-guard without dependency loops.
+        "//include/envoy/runtime:runtime_interface",
         "//source/common/common:hash_lib",
         "//source/common/singleton:const_singleton",
+        "//source/common/singleton:threadsafe_singleton",
     ],
 )
+
+envoy_cc_library(
+    name = "runtime_protos_lib",
+    hdrs = [
+        "runtime_protos.h",
+    ],
+    deps = [
+        "//include/envoy/runtime:runtime_interface",
+        "@envoy_api//envoy/service/runtime/v3:pkg_cc_proto",
+    ]
+    )
 
 envoy_cc_library(
     name = "runtime_lib",
@@ -29,11 +45,11 @@ envoy_cc_library(
     ],
     hdrs = [
         "runtime_impl.h",
-        "runtime_protos.h",
     ],
     external_deps = ["ssl"],
     deps = [
         ":runtime_features_lib",
+        ":runtime_protos_lib",
         "//include/envoy/config:subscription_interface",
         "//include/envoy/event:dispatcher_interface",
         "//include/envoy/init:manager_interface",
