+++ source/common/runtime/runtime_features.h
@@ -2,12 +2,19 @@
 
 #include <string>
 
+#include "envoy/runtime/runtime.h"
+
 #include "common/singleton/const_singleton.h"
+#include "common/singleton/threadsafe_singleton.h"
 
 #include "absl/container/flat_hash_set.h"
 
 namespace Envoy {
 namespace Runtime {
+
+bool isRuntimeFeature(absl::string_view feature);
+bool runtimeFeatureEnabled(absl::string_view feature);
+uint64_t getInteger(absl::string_view feature, uint64_t default_value);
 
 class RuntimeFeatures {
 public:
