+++ source/common/runtime/runtime_impl.h
@@ -32,9 +32,6 @@
 namespace Envoy {
 namespace Runtime {
 
-bool runtimeFeatureEnabled(absl::string_view feature);
-uint64_t getInteger(absl::string_view feature, uint64_t default_value);
-
 using RuntimeSingleton = ThreadSafeSingleton<Loader>;
 
 /**
