+++ test/common/runtime/runtime_flag_override_test.cc
@@ -1,4 +1,4 @@
-#include "common/runtime/runtime_impl.h"
+#include "common/runtime/runtime_features.h"
 
 #include "gmock/gmock.h"
